Support multi-push OP_RETURN scripts
Inspired by @petertodd's multi-push branch:

 - https://github.com/petertodd/bitcoin/tree/op-return-accept-multiple-pushdatas

This goes hand in hand with being able to handle multiple OP_RETURN outputs in one transaction, with any size, despite user defined settings.

Example:

A transaction with the following outputs:

 - OP_RETURN PUSH 0x6f6d PUSH 0x00000000000000010000000006dac2c0 PUSH 05627138bb55251bfb289a1ec390eafd3755b1a698
 - OP_RETURN PUSH 0x6f6d PUSH 0x00000000000000070000000000002329 PUSH 0062e907b15cbf27d5425399ebf6f0fb50ebb88f18

... yields the following (already combined and stripped) payload:

 - 00000000000000010000000006dac2c005627138bb55251bfb289a1ec390eafd3755b1a698000000000000000700000000000023290062e907b15cbf27d5425399ebf6f0fb50ebb88f18

It does not affect the behavior of Bitcoin Core, but only the meta trasaction processing, which is independent and shielded from "standardness".