Add 0.0.9 parseTransaction as legacy code
Legacy code is kept to guarantee a safe transition until a new feature, or a replacement of consensus critical code is enabled.

Once the new replacement code is activated on mainnet, and it is conform with historical consensus, the legacy code can safely be removed.

The adjustments were kept as minimal as possible.