As per default, don't log packet parsing in RPC mode
The debug flag "packets_readonly" is used to enable subsequent logging of packet parsing during RPC calls and alike, and is disabled as per default, so that only the first, actual, packet interpretation is logged.

This commit further introduces a flag to indicate, whether a transaction is used to execute logic. To use, or access, "interpretPacket()", "unlockLogic()" must be called first. It serves as workaround to indicate, when packet parsing should be logged, but also as guard, to prevent accidental, or not explicitly requested, execution of logic.