PoC: REST endpoint for "omni_gettransaction"
Only JSON output is supported.

To test, add `rest=1` to `bitcoin.conf`, or start with `./src/omnicored -rest`.

Assuming a default setup, with standard RPC port 8332, then transactions are avaiable via:

  http://localhost:8332/omni/tx/:hash.json

Example (mainnet):

  http://localhost:8332/rest/omni/tx/3ea6fcea096107c01ce4972e849200515beafcae9aa7d550309f466199574124.json

The endpoint can be queried via a browser, curl, and friends.

Note: the API is public and not authenticated, i.e. no credentials are required to access the REST interface.